+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # CHANGELOG
 
+## Version 1.0.3
+
+- Refactor test:
+  - Reorganize the structure
+  - Improve coverage
+  - Comply with rubocop
+- Refactor and fix lib:
+  - Holdify is a module
+  - Simplify find_location
+  - Simplify and improve holdify_plugin.rb
+
 ## Version 1.0.2
 
 - Rename rebuild > reconcile; improve README.md